Refactored async related gameManager methods.

+++ Assets/Scripts/NonUnityDependent/UnityChessLib/Base/Timeline.cs
@@ -3,25 +3,25 @@ using System.Collections.Generic;
 
 namespace UnityChess {
 	public class Timeline<T> {
-		public T Current => list[headIndex];
+		public T Current => list[headIndexBacking];
 		public int Span => list.Count;
-		public bool IsUpToDate => headIndex == list.Count - 1;
+		public bool IsUpToDate => headIndexBacking == list.Count - 1;
 		public int HeadIndex {
-			get => headIndex;
-			set => headIndex = Math.Min(value, list.Count - 1);
-		}
-
-		private int headIndex;
+			get => headIndexBacking;
+			set => headIndexBacking = Math.Min(value, list.Count - 1);
+		} private int headIndexBacking;
+		
+		
 		private readonly List<T> list;
-		private int FutureElementsStartIndex => headIndex + 1;
+		private int FutureElementsStartIndex => headIndexBacking + 1;
 		private int NumFutureElements => list.Count - FutureElementsStartIndex;
 
 		public Timeline() {
-			headIndex = -1;
+			headIndexBacking = -1;
 			list = new List<T>();
 		}
 		
-		public List<T> GetStartToCurrent() => list.GetRange(0, headIndex + 1);
+		public List<T> GetStartToCurrent() => list.GetRange(0, headIndexBacking + 1);
 
 		public List<T> PopFuture() {
 			List<T> elementRange = list.GetRange(FutureElementsStartIndex, NumFutureElements);
@@ -32,17 +32,16 @@ namespace UnityChess {
 		public void AddNext(T element) {
 			Prune();
 			list.Add(element);
-			headIndex++;
+			headIndexBacking++;
 		}
 
 		private void Prune() {
-			if (IsUpToDate) return;
-			list.RemoveRange(FutureElementsStartIndex, NumFutureElements);
+			if (!IsUpToDate) list.RemoveRange(FutureElementsStartIndex, NumFutureElements);
 		}
 		
 		public void Clear() {
 			list.Clear();
-			headIndex = -1;
+			headIndexBacking = -1;
 		}
 	}
 }
